+++ servers/slapd/back-shell/modrdn.c
@@ -52,7 +52,7 @@ shell_back_modrdn(
 
 	/* write out the request to the modrdn process */
 	fprintf( wfp, "MODRDN\n" );
-	fprintf( wfp, "msgid: %ld\n", op->o_msgid );
+	fprintf( wfp, "msgid: %ld\n", (long) op->o_msgid );
 	print_suffixes( wfp, be );
 	fprintf( wfp, "dn: %s\n", dn );
 	fprintf( wfp, "newrdn: %s\n", newrdn );
